refactor: split work cli family
