+++ src/brigade/cli/work/__init__.py
@@ -1,0 +1,72 @@
+"""Compatibility facade for the split command family."""
+# ruff: noqa: E402,F401,F403,F811,F821
+
+from __future__ import annotations
+
+import importlib
+import sys
+from types import ModuleType
+from typing import Any
+
+_MODULE_NAMES = (
+    "register",
+    "dispatch",
+    "phases",
+)
+
+
+def __getattr__(name: str) -> Any:
+    if name in _MODULE_NAMES:
+        module = importlib.import_module(f"{__name__}.{name}")
+        globals()[name] = module
+        return module
+    for module_name in _MODULE_NAMES:
+        module = importlib.import_module(f"{__name__}.{module_name}")
+        if hasattr(module, name):
+            value = getattr(module, name)
+            globals()[name] = value
+            return value
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+
+_module_register = importlib.import_module(f"{__name__}.register")
+_module_dispatch = importlib.import_module(f"{__name__}.dispatch")
+_module_phases = importlib.import_module(f"{__name__}.phases")
+
+from .register import *
+from .dispatch import *
+from .phases import *
+
+_MODULES = (
+    _module_register,
+    _module_dispatch,
+    _module_phases,
+)
+
+
+def _sync_module_globals() -> None:
+    exported: dict[str, Any] = {}
+    for module in _MODULES:
+        for name, value in vars(module).items():
+            if name.startswith("__"):
+                continue
+            exported.setdefault(name, value)
+    facade = sys.modules[__name__]
+    for name, value in exported.items():
+        setattr(facade, name, value)
+    for module in _MODULES:
+        for name, value in exported.items():
+            if name not in module.__dict__:
+                setattr(module, name, value)
+
+
+class _CommandFamilyFacade(ModuleType):
+    def __setattr__(self, name: str, value: Any) -> None:
+        super().__setattr__(name, value)
+        for module in _MODULES:
+            if hasattr(module, name):
+                setattr(module, name, value)
+
+
+_sync_module_globals()
+sys.modules[__name__].__class__ = _CommandFamilyFacade
